+++ src/components/templates/LoginOnly.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { navigate } from "gatsby";
 import { useSelector } from "react-redux";
-import { Router } from "@reach/router";
 
 export const LoginOnly: React.FC = ({ children }) => {
   const { userDoc, loading } = useSelector((state) => state.auth);
@@ -16,5 +15,5 @@ export const LoginOnly: React.FC = ({ children }) => {
     return null;
   }
 
-  return <Router>{children}</Router>;
+  return <>{children}</>;
 };
