+++ src/pages/index.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Router } from "@reach/router";
 
 import { Layout } from "../components/templates/Layout";
 import { Index } from "../components/pages/Index";
@@ -7,9 +6,7 @@ import { Index } from "../components/pages/Index";
 const IndexPage: React.FC = () => {
   return (
     <Layout>
-      <Router>
-        <Index path="/" />
-      </Router>
+      <Index />
     </Layout>
   );
 };
